requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
requests==2.28.1
PyJWT==2.4.0
python-dotenv==0.20.0
cryptography==42.0.8
flask-restx==1.3.0
flask-swagger-ui==4.11.1
PyYAML==6.0.1
hl7==0.4.5
black==23.7.0
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-watch==4.2.0



